killms debugpdb off and fiddle with separators
